Split lane_markings.rs into subdirectory with focused modules
Break the 771-line lane_markings.rs into four smaller files under
lane_markings/:

- mod.rs (175 lines): components, constants, and the sync system
- mesh_builder.rs (212 lines): per-segment mesh construction
- line_primitives.rs (187 lines): Bezier helpers and line emitters
- tests.rs (222 lines): all unit tests

All pub items re-exported from mod.rs so external imports are unchanged.

Part of #1389
